Move tests with common setup together to speed up testing time
