+++ tests/integration/security/citadel_namespace_targeting_test.go
