+++ tests/integration/security/main_test.go
@@ -12,7 +12,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package manualmtls
+package security
 
 import (
 	"testing"
@@ -33,7 +33,7 @@ var (
 
 func TestMain(m *testing.M) {
 	framework.
-		NewSuite("reachability_test", m).
+		NewSuite("security", m).
 		SetupOnEnv(environment.Kube, istio.Setup(&ist, nil)).
 		Setup(func(ctx resource.Context) (err error) {
 			if g, err = galley.New(ctx, galley.Config{}); err != nil {
