+++ tests/integration/security/testdata/citadel-noread-rbac.yaml.tmpl
