+++ tests/integration/security/testdata/citadel-rbac.yaml.tmpl
