+++ tests/integration/security/testdata/global-mtls-on-no-dr.yaml
