+++ tests/integration/security/testdata/global-mtls-permissive.yaml
