+++ tests/integration/security/testdata/jwt-with-paths.yaml.tmpl
