+++ tests/integration/security/testdata/simple-jwt-policy.yaml.tmpl
